Add Virtual Kubelet
* Announced at the Azure/Docker/Kubernetes meetup in Seattle today.
Virtual Kubelet is an open source Kubernetes kubelet implementation that masquerades as a kubelet for the purposes of connecting Kubernetes to other APIs. This allows the nodes to be backed by other services like ACI, Hyper.sh, AWS, Azure, etc. This connector features a pluggable architecture and direct use of Kubernetes primitives, making it much easier to build on.