Remove Fabric8 and Openshift.io
Fabric8 upstream project is abandoned and redhat's fork of it called openshift.io is not ready for public consumption yet as it's still a private beta. Both do not add any value to this otherwise pretty great list of resources.  

Source: I used to work on these projects at Redhat.